rename event_emit folder to event-emit
